Extract passport and session middleware
